Add a proper download_all function and improve logging
Formerly, we used Python's all function to download all the search
results. While it does work because searcher _is_ iterable, I do
not really like that design.

The all function does loops through all the items but it is designed
to check if everythin in an iterable is True. As a side effect of
this behavior, we ended up downloading everything because the
access of an item causes it to be downloaded. I don't think that
twisting the all function this way makes sense.